restructure folders for pass debrief page for cat-be
