+++ src/pages/test-finalisation/pass-finalisation/pass-finalisation.module.ts
@@ -1,20 +1,13 @@
 import { NgModule } from '@angular/core';
-import { IonicPageModule } from 'ionic-angular';
-import { PassFinalisationPage } from './pass-finalisation';
-import { EffectsModule } from '@ngrx/effects';
-import { PassFinalisationAnalyticsEffects } from './pass-finalisation.analytics.effects';
-import { ComponentsModule } from '../../../components/common/common-components.module';
-import { TestFinalisationComponentsModule } from '../components/test-finalisation.module';
+import { PassFinalisationCatBPageModule } from './cat-b/pass-finalisation.module.cat-b.page';
+
 
 @NgModule({
-  declarations: [
-    PassFinalisationPage,
-  ],
+
   imports: [
-    IonicPageModule.forChild(PassFinalisationPage),
-    EffectsModule.forFeature([PassFinalisationAnalyticsEffects]),
-    ComponentsModule,
-    TestFinalisationComponentsModule,
+    PassFinalisationCatBPageModule
   ],
 })
+
 export class PassFinalisationPageModule {}
+
